Use /Courses instead of legacy EducationalEvents #1099
