+++ src/app/edit-absences/components/edit-absences-header/edit-absences-header.component.html
@@ -16,8 +16,8 @@
     }}</label>
     <bkd-typeahead
       [id]="'filter-module-instance'"
-      [typeaheadService]="educationalEventsService"
-      [(value)]="filter.educationalEvent"
+      [typeaheadService]="coursesService"
+      [(value)]="filter.course"
     ></bkd-typeahead>
   </div>
